Allow initrd to be empty
The objtools is unhappy if the initrd is either /dev/null or an empty
file.  Check, in the Makefile, for /dev/null, and if so, create a
small empty file for the initrd.  The kernel will ignore an initrd
that doesn't have the proper header, so this will achieve the same
effect.

Signed-off-by: David Brown <david.brown@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>